requirements.txt
pyyaml
adafruit-circuitpython-servokit
flask
requests
opencv-python
pyzmq